Add test to verify that moving page to unsaved structure in gallery works
